requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
streamlit
openai
python-dotenv

#api
langserve
fastapi
uvicorn

#rag
bs4
pypdf
chromadb==0.3.29
faiss-cpu


#agents
arxiv
wikipedia
langchainhub
youtube_search



#rag with groq
groq
beautifulsoup4
langchain-groq

#mistral
sentence_transformers


#object box
langchain-objectbox